Added 'trace' feature to vergen-fmt, some refactoring as a result
