+++ .github/workflows/main.yml
@@ -153,15 +153,15 @@ jobs:
       - name: 💾 Install (tarpaulin, all-features) 💾
         run: cargo binstall --no-confirm --no-symlinks cargo-tarpaulin cargo-all-features
       - name: 🧪 Test 🧪
-        run: cargo test-all-features
-      - name: 🧪 Test gitcl 🧪
-        run: cargo test -p vergen -F build,cargo,git,gitcl,rustc,si
-      - name: 🧪 Test git2 🧪
-        run: cargo test -p vergen -F build,cargo,git,git2,rustc,si
-      - name: 🧪 Test gix 🧪
-        run: cargo test -p vergen -F build,cargo,git,gix,rustc,si
-      - name: 🧪 Test vergen-fmt 🧪
-        run: cargo test -p vergen-fmt -F __vergen_test
+        run: |
+          cargo test-all-features
+          cargo test -p vergen -F build,cargo,git,gitcl,rustc,si
+          cargo test -p vergen -F build,cargo,git,git2,rustc,si
+          cargo test -p vergen -F build,cargo,git,gix,rustc,si
+          cargo test -p vergen-fmt -F __vergen_test
+          cargo test -p vergen-fmt -F __vergen_test,color
+          cargo test -p vergen-fmt -F __vergen_test,trace
+          cargo test -p vergen-fmt -F __vergen_test,color,trace
       - name: 🙊 Code Coverage (vergen) 🙊
         run: cargo tarpaulin --config vergen/tarpaulin.toml --exclude-files vergen/src/lib.rs --exclude-files vergen-fmt/* --out Xml --output-dir vergen
       - name: 🙊 Code Coverage (vergen-fmt) 🙊
@@ -212,15 +212,15 @@ jobs:
       - name: 💾 Install (tarpaulin, all-features) 💾
         run: cargo binstall --no-confirm --no-symlinks cargo-all-features
       - name: 🧪 Test 🧪
-        run: cargo test-all-features
-      - name: 🧪 Test gitcl 🧪
-        run: cargo test -p vergen -F build,cargo,git,gitcl,rustc,si
-      - name: 🧪 Test git2 🧪
-        run: cargo test -p vergen -F build,cargo,git,git2,rustc,si
-      - name: 🧪 Test gix 🧪
-        run: cargo test -p vergen -F build,cargo,git,gix,rustc,si
-      - name: 🧪 Test vergen-fmt 🧪
-        run: cargo test -p vergen-fmt -F __vergen_test
+        run: |
+          cargo test-all-features
+          cargo test -p vergen -F build,cargo,git,gitcl,rustc,si
+          cargo test -p vergen -F build,cargo,git,git2,rustc,si
+          cargo test -p vergen -F build,cargo,git,gix,rustc,si
+          cargo test -p vergen-fmt -F __vergen_test
+          cargo test -p vergen-fmt -F __vergen_test,color
+          cargo test -p vergen-fmt -F __vergen_test,trace
+          cargo test -p vergen-fmt -F __vergen_test,color,trace
   test-windows:
     name: 🧪 Windows 🧪
     needs: lints-windows
@@ -275,12 +275,12 @@ jobs:
       - name: 💾 Install (all-features) 💾
         run: cargo binstall --no-confirm --no-symlinks cargo-all-features
       - name: 🧪 Test 🧪
-        run: cargo test-all-features
-      - name: 🧪 Test gitcl 🧪
-        run: cargo test -p vergen -F build,cargo,git,gitcl,rustc,si
-      - name: 🧪 Test git2 🧪
-        run: cargo test -p vergen -F build,cargo,git,git2,rustc,si
-      - name: 🧪 Test gix 🧪
-        run: cargo test -p vergen -F build,cargo,git,gix,rustc,si
-      - name: 🧪 Test vergen-fmt 🧪
-        run: cargo test -p vergen-fmt -F __vergen_test
+        run: |
+          cargo test-all-features
+          cargo test -p vergen -F build,cargo,git,gitcl,rustc,si
+          cargo test -p vergen -F build,cargo,git,git2,rustc,si
+          cargo test -p vergen -F build,cargo,git,gix,rustc,si
+          cargo test -p vergen-fmt -F __vergen_test
+          cargo test -p vergen-fmt -F __vergen_test,color
+          cargo test -p vergen-fmt -F __vergen_test,trace
+          cargo test -p vergen-fmt -F __vergen_test,color,trace
