+++ run_all.fish
@@ -10,5 +10,8 @@ cargo test -p vergen -F build,cargo,git,gitcl,rustc,si; and \
 cargo test -p vergen -F build,cargo,git,git2,rustc,si; and \
 cargo test -p vergen -F build,cargo,git,gix,rustc,si; and \
 cargo test -p vergen-fmt -F __vergen_test; and \
+cargo test -p vergen-fmt -F __vergen_test,color; and \
+cargo test -p vergen-fmt -F __vergen_test,trace; and \
+cargo test -p vergen-fmt -F __vergen_test,color,trace; and \
 cargo doc -p vergen -F build,cargo,git,gitcl,rustc,si; and \
 cargo doc -p vergen-fmt
