+++ vergen-fmt/Cargo.toml
@@ -15,12 +15,14 @@ version = "0.1.0"
 [package.metadata.cargo-all-features]
 denylist = [
     "console",
+    "tracing",
     "vergen",
 ]
 
 [features]
 default = []
 color = ["console"]
+trace = ["tracing"]
 __vergen_test = ["vergen"]
 
 [dependencies]
@@ -28,6 +30,7 @@ anyhow = "1.0.68"
 console = { version = "0.15.4", optional = true }
 indexmap = "1.9.2"
 lazy_static = "1.4.0"
+tracing = { version = "0.1.37", features = ["max_level_trace", "release_max_level_trace"], optional = true }
 
 [build-dependencies]
 anyhow = "1.0.68"
